Determines the number of bits needed to be flipped to get from one number to another

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -17,7 +17,7 @@
   */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int new, count;
+	unsigned long int new, count;
 
 	new = n ^ m;
 	count = 0;
